Move some SRv6 functions to srv6_helper.py
